requirements.txt
lifelines==0.27.8
pandas==1.5.3
plotly==5.9.0
scipy==1.10.0
streamlit==1.37.1
streamlit_plotly_events==0.0.6
openpyxl==3.1.5
streamlit-aggrid==1.0.5
numpy
pysam
pybedtools
pyyaml
ipykernel

